Implement the health check API as a protocol agnostic `core` module
Move the underlying implementation of the `_health` HTTP API to the
`core` module and update the HTTP method to call the core impl. Also,
add support for invoking the health check API from the CLI using the
same underlying core impl.

Closes https://github.com/roadster-rs/roadster/issues/170
