+++ src/api/core/mod.rs
@@ -1,0 +1,3 @@
+//! Internal, protocol agnostic API implementations. The APIs/methods provided here provide the
+//! underlying implementation for the HTTP, gPRC, CLI, etc APIs.
+pub mod health;
