+++ src/api/mod.rs
@@ -1,4 +1,5 @@
 #[cfg(feature = "cli")]
 pub mod cli;
+pub mod core;
 #[cfg(feature = "http")]
 pub mod http;
